Исправления отрисовки полного изображения
