+++ js/create-message.js
@@ -1,0 +1,8 @@
+const createMessage = function (element) {
+  const text = document.createElement('p');
+  text.classList.add('social__text');
+  text.textContent = element.message;
+  return text;
+};
+
+export {createMessage};
