+++ js/create-picture.js
@@ -1,6 +1,6 @@
 import {openBigPictureModal} from './big-picture-modal.js';
 
-const createPicture = (pictures) => {
+const createPicture = function (pictures) {
   const pictureListElement = document.querySelector('.pictures');
   const pictureTemplate = document.querySelector('#picture').content;
   const pictureListFragment = document.createDocumentFragment();
@@ -8,7 +8,7 @@ const createPicture = (pictures) => {
   pictures.forEach((picture, index) => {
     const pictureElement = pictureTemplate.cloneNode(true);
     pictureElement.querySelector('.picture__img').src = picture.url;
-    pictureElement.querySelector('.picture__img').setAttribute('data-img-id', index);
+    pictureElement.querySelector('.picture__img').dataset.imgId = index;
     pictureElement.querySelector('.picture__likes').textContent = picture.likes;
     pictureElement.querySelector('.picture__comments').textContent = picture.comments.length;
     pictureListFragment.appendChild(pictureElement);
